+++ ocaml-lsp-server/src/ocamlformat.mli
@@ -12,4 +12,5 @@ type error =
 
 val message : error -> string
 
-val run : Document.t -> (TextEdit.t list, error) result Fiber.t
+val run :
+  Document.t -> Fiber.Cancel.t option -> (TextEdit.t list, error) result Fiber.t